file structure and declaration files
