+++ .prettierignore
@@ -1,0 +1,4 @@
+.eslintrc
+tsconfig.json
+webpack.config.js
+types
