+++ types/index.d.ts
@@ -1,0 +1,3 @@
+export * from './services/side-panels/types/types';
+export { SidePanelsManager } from './services/side-panels/side-panels-manager';
+export { ISidePanelItemDto } from './services/side-panels/models/side-panel-item-dto';
