+++ types/services-managers.d.ts
@@ -1,0 +1,1 @@
+export declare const pluginName = "uiManagers";
