+++ types/services/side-panels/models/item-wrapper.d.ts
@@ -1,0 +1,1 @@
+export {};
