+++ types/services/side-panels/types/types.d.ts
@@ -1,0 +1,3 @@
+export declare type SidePanelPosition = 'top' | 'bottom' | 'right' | 'left';
+export declare type SidePanelMode = 'alongside' | 'hidden' | 'over';
+export declare type ReservedPresetName = 'Playback' | 'Live';
